chapterup/requirements.txt
docopt-ng
imgur_python
natsort
tqdm
xdg
